client/: Merge client.rs into thin_client.rs
