+++ client/src/thin_client.rs
@@ -202,6 +202,20 @@ impl Drop for ThinClient {
     }
 }
 
+pub fn create_client((rpc, tpu): (SocketAddr, SocketAddr), range: (u16, u16)) -> ThinClient {
+    let (_, transactions_socket) = solana_netutil::bind_in_range(range).unwrap();
+    ThinClient::new(rpc, tpu, transactions_socket)
+}
+
+pub fn create_client_with_timeout(
+    (rpc, tpu): (SocketAddr, SocketAddr),
+    range: (u16, u16),
+    timeout: Duration,
+) -> ThinClient {
+    let (_, transactions_socket) = solana_netutil::bind_in_range(range).unwrap();
+    ThinClient::new_socket_with_timeout(rpc, tpu, transactions_socket, timeout)
+}
+
 pub fn retry_get_balance(
     client: &ThinClient,
     bob_pubkey: &Pubkey,
